requirements.txt
discord.py
google-generativeai
flask
python-dotenv
duckduckgo-search